IPlugAPP: Explict cast to fix conversion warning

+++ IPlug/APP/IPlugAPP.cpp
@@ -62,10 +62,11 @@ bool IPlugAPP::EditorResize(int viewWidth, int viewHeight)
     float ss = 1.f;
     #endif
     
-    SetWindowPos(gHWND, 0, rcWindow.left * ss,
-                 (rcWindow.bottom - viewHeight - ptDiff.y) * ss,
-                 (viewWidth + ptDiff.x) * ss,
-                 (viewHeight + ptDiff.y) * ss, flags);
+    SetWindowPos(gHWND, 0,
+                 static_cast<LONG>(rcWindow.left * ss),
+                 static_cast<LONG>((rcWindow.bottom - viewHeight - ptDiff.y) * ss),
+                 static_cast<LONG>((viewWidth + ptDiff.x) * ss),
+                 static_cast<LONG>((viewHeight + ptDiff.y) * ss), flags);
     parentResized = true;
     #endif
